% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_noise.R
\docType{data}
\name{data_noise}
\alias{data_noise}
\title{data_noise}
\format{
A data frame with 2000 columns across 5000 rows \cr
Columns correspond to cells and rows to genes. \cr
}
\usage{
data_noise
}
\description{
An artificial data set used to derive different sparsity level
}
\keyword{datasets}
