% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sparsity.R
\name{get_sparsity}
\alias{get_sparsity}
\title{Retrieve percentage of zeros from data set}
\usage{
get_sparsity(data_input)
}
\arguments{
\item{data_input}{data frame, the data from which the percentage of zeros should be calculated}
}
\value{
num, The calculated percentage of zeros in the data set
}
\description{
Retrieve percentage of zeros from data set
}
\examples{
# use gold data 
get_sparsity(sc_gold)

}
