% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce_sparsity.R
\name{introduce_sparsity}
\alias{introduce_sparsity}
\title{Generate a sparse data set from a reference}
\usage{
introduce_sparsity(lamdba_index)
}
\arguments{
\item{lamdba_index}{numeric, range from 0 to 1, \cr
close to zero results in a more sparse data set}
}
\value{
A dataframe of sparsified data
}
\description{
Generate a sparse data set from a reference
}
\examples{
# data(sc_gold)
sparse_data <- introduce_sparsity(0.1)

}
