% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_gold.R
\docType{data}
\name{sc_gold}
\alias{sc_gold}
\title{sc_gold}
\format{
A data frame with 800 rows across 4960 columns. Rows correspond to cells and columns to genes. The data set was already filtered for missing entries.
}
\usage{
sc_gold
}
\description{
Am artificial, non-sparse scRNA-seq reference data set
}
\keyword{datasets}
