% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_distribution.R
\name{Plot_distribution}
\alias{Plot_distribution}
\title{Plot distribution of expression values}
\usage{
Plot_distribution(data)
}
\arguments{
\item{data}{A data frame used to plot its distribution}
}
\value{
A density plot of expression values
}
\description{
Plot distribution of expression values
}
\examples{
# use gold data
Plot_distribution(data = sc_gold)
}
